Cleaned up code. Made new component for edit form and renamed EditMedicine component to ManageInventory.jsx
